+++ src/boot/register-components.js
@@ -3,7 +3,9 @@ import AppMenu from '../components/AppMenu.vue'
 import AppsMenu from '../components/AppsMenu.vue'
 import Autocomplete from '../components/Autocomplete.vue'
 import Avatar from '../components/Avatar.vue'
+import Box from '../components/Box.vue'
 import BreakLine from '../components/BreakLine.vue'
+import Btn from '../components/Btn.vue'
 import CheckboxGroup from '../components/CheckboxGroup.vue'
 import Copy from '../components/Copy.vue'
 import DatetimeInput from '../components/DatetimeInput.vue'
@@ -15,6 +17,7 @@ import Field from '../components/Field.vue'
 import Filters from '../components/Filters.vue'
 import FormGenerator from '../components/FormGenerator.vue'
 import FormView from '../components/FormView.vue'
+import GridGenerator from '../components/GridGenerator.vue'
 import Input from '../components/Input.vue'
 import ListView from '../components/ListView.vue'
 import PageHeader from '../components/PageHeader.vue'
@@ -22,13 +25,14 @@ import PasswordInput from '../components/PasswordInput.vue'
 import PasswordStrengthChecker from '../components/PasswordStrengthChecker.vue'
 import Profile from '../components/Profile.vue'
 import Resizer from '../components/Resizer.vue'
+import SearchBox from '../components/SearchBox.vue'
+import SelectList from '../components/SelectList.vue'
+import SettingsMenu from '../components/SettingsMenu.vue'
 import SingleView from '../components/SingleView.vue'
 import Sortable from '../components/Sortable.vue'
 import TableGenerator from '../components/TableGenerator.vue'
 import TabsGenerator from '../components/TabsGenerator.vue'
 import Uploader from '../components/Uploader.vue'
-import Box from '../components/Box.vue'
-import SearchBox from '../components/SearchBox.vue'
 
 export default async ({ Vue }) => {
   Vue.component('qs-app-bar', AppBar)
@@ -38,6 +42,7 @@ export default async ({ Vue }) => {
   Vue.component('qs-avatar', Avatar)
   Vue.component('qs-box', Box)
   Vue.component('qs-break-line', BreakLine)
+  Vue.component('qs-btn', Btn)
   Vue.component('qs-checkbox-group', CheckboxGroup)
   Vue.component('qs-copy', Copy)
   Vue.component('qs-datetime-input', DatetimeInput)
@@ -49,6 +54,7 @@ export default async ({ Vue }) => {
   Vue.component('qs-filters', Filters)
   Vue.component('qs-form-generator', FormGenerator)
   Vue.component('qs-form-view', FormView)
+  Vue.component('qs-grid-generator', GridGenerator)
   Vue.component('qs-input', Input)
   Vue.component('qs-list-view', ListView)
   Vue.component('qs-page-header', PageHeader)
@@ -57,6 +63,8 @@ export default async ({ Vue }) => {
   Vue.component('qs-profile', Profile)
   Vue.component('qs-resizer', Resizer)
   Vue.component('qs-search-box', SearchBox)
+  Vue.component('qs-select-list', SelectList)
+  Vue.component('qs-settings-menu', SettingsMenu)
   Vue.component('qs-single-view', SingleView)
   Vue.component('qs-sortable', Sortable)
   Vue.component('qs-table-generator', TableGenerator)